refactor: re-arrange existing resources components to support different attributes mappings
