+++ plugin-hrm-form/src/components/resources/mappingComponents/ExpandableAttributeContent.tsx
