+++ plugin-hrm-form/src/components/resources/mappingComponents/khpMappings/ExpandableSection.tsx
@@ -17,8 +17,8 @@ import React, { useState } from 'react';
 import ArrowRightTwoTone from '@material-ui/icons/ArrowRightTwoTone';
 import ArrowDropDownTwoTone from '@material-ui/icons/ArrowDropDownTwoTone';
 
-import { SectionCollapse } from '../../search/styles';
-import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../styles';
+import { SectionCollapse } from '../../../search/styles';
+import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../../styles';
 
 type ExpandableSectionProps = {
   title: string | JSX.Element;
