+++ plugin-hrm-form/src/components/resources/mappingComponents/khpMappings/SiteDetails.tsx
@@ -16,13 +16,13 @@
 
 import React from 'react';
 
-import { KhpUiResource } from '../types';
-import { FontOpenSans } from '../../../styles';
-import { ResourceAttributeContent, ResourceAttributeDescription, ResourceSubtitle } from '../styles';
+import { KhpUiResource } from './types';
+import { FontOpenSans } from '../../../../styles';
+import { ResourceAttributeContent, ResourceAttributeDescription, ResourceSubtitle } from '../../styles';
 import ExpandableSection from './ExpandableSection';
 import OperatingHours from './OperatingHours';
-import ResourceAttributeWithPrivacy from './ResourceAttributeWithPrivacy';
-import ResourceAttribute from './ResourceAttribute';
+import ResourceAttributeWithPrivacy from '../ResourceAttributeWithPrivacy';
+import ResourceAttribute from '../ResourceAttribute';
 
 type Props = {
   sites: KhpUiResource['attributes']['site'];
